+++ src/main/java/com/atlassian/db/replica/api/DualConnection.java
@@ -61,7 +61,8 @@ public final class DualConnection implements Connection {
         this.connectionProvider = new ReplicaConnectionProvider(
             connectionProvider,
             consistency,
-            stateListener
+            stateListener,
+            compatibleWithPreviousVersion
         );
         this.consistency = consistency;
         this.databaseCall = databaseCall;