+++ CHANGELOG.md
@@ -24,6 +24,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-2.1.4...master
 
+### Added
+- A new reason REPLICA_NOT_AVAILABLE
+
 ## [2.1.4] - 2021-07-08
 [2.1.4]: https://github.com/atlassian-labs/db-replica/compare/release-2.1.2...release-2.1.4